remove not used classes (moved them)
